requirements.txt
beautifulsoup4
pika
pymongo
python-dateutil
python-jsonrpc
redis
requests
six